tests/qaudio: Add Test Case to play a wav file.
This test only plays a a wav file, a user must confirm that the file is played.

+++ igep_qa/tests/qaudio.py
@@ -15,17 +15,6 @@ class TestAudio(unittest.TestCase):
     """ Generic tests for audio interfaces.
 
     """
-    def setUp(self):
-        """ Ensure requirements are installed.
-
-        """
-        required = ["aplay", "arecord", "multimon"]
-        files = "/usr/share/igep_qa/contrib/dtmf.wav"
-        for req in required:
-            if not is_in_path(req):
-                raise Exception("Can't find %s" % req)
-        if not os.path.isfile(files):
-            raise Exception("Can't find %s" % files)
 
     def test_audio_loopback(self):
         """ Test Audio : Loopback, sound sent to audio-out should return in audio-in
@@ -51,6 +40,14 @@ class TestAudio(unittest.TestCase):
               gets the DMTF digits.
 
         """
+        # Ensure requirements are installed.
+        required = ["aplay", "arecord", "multimon"]
+        files = "/usr/share/igep_qa/contrib/dtmf.wav"
+        for req in required:
+            if not is_in_path(req):
+                raise Exception("Can't find %s" % req)
+        if not os.path.isfile(files):
+            raise Exception("Can't find %s" % files)
         commands.getoutput("rm -f /tmp/recorded.wav")
         commands.getoutput("aplay -t wav -v /usr/share/igep_qa/contrib/dtmf.wav & arecord -t wav -c 1 "
                            "-r 8000 -f S16_LE -d 5 -v /tmp/recorded.wav")
@@ -60,6 +57,33 @@ class TestAudio(unittest.TestCase):
         
         self.failUnless(retval[0] == 0, "failed: No DTMF found in recorded "
                         "file")
+
+    def test_audio_playwav(self):
+        """ Test Audio : Play a wav file
+
+        Type: Functional
+
+        Prerequisite commands:
+            - aplay
+
+        Prerequitsite files:
+            - test.wav (on contrib/test.wav)
+
+        Description:
+            - This test only plays a wav file, a user must confirm that the
+            file is played.
+
+        """
+        # Ensure requirements are installed.
+        required = ["aplay"]
+        files = "/usr/share/igep_qa/contrib/test.wav"
+        for req in required:
+            if not is_in_path(req):
+                raise Exception("Can't find %s" % req)
+        if not os.path.isfile(files):
+            raise Exception("Can't find %s" % files)
+        retval = commands.getstatusoutput("aplay -t wav /usr/share/igep_qa/contrib/test.wav")
+        self.failUnless(retval[0] == 0, "failed: Playing test.wav.")
 
 if __name__ == '__main__':
     unittest.main(verbosity=2)
